Added new Communication category with comms library.
The comms library is the C++(11) headers only, platform independent one, which makes the
implementation of a communication protocol to be an easy and relatively quick process.
It provides all the necessary types and classes to make the definition of the custom messages,
as well as wrapping transport data fields, to be simple declarative statements of type and class definitions.